refactor: update paths
