+++ src/features/home/contribute/contribute.tsx
@@ -1,7 +1,7 @@
 import { OptionItem, type OptionItemProps } from '@/app/components';
 
-import image from '@/assets/contribute.webp';
 import Image from '@/features/image';
+import image from '@/shared/assets/contribute.webp';
 import { Paragraph } from '@/shared/ui/paragraph';
 import { Subtitle } from '@/shared/ui/subtitle';
 import { Title } from '@/shared/ui/title';
