+++ src/features/mentors-wanted/mentors-wanted.tsx
@@ -1,8 +1,8 @@
 import classNames from 'classnames';
 import classNamesBind from 'classnames/bind';
 import { Link } from 'react-router-dom';
-import mentorImg from '@/assets/mentors-wanted-poster.webp';
 import Image from '@/features/image';
+import mentorImg from '@/shared/assets/mentors-wanted-poster.webp';
 import { Paragraph } from '@/shared/ui/paragraph';
 import { Title } from '@/shared/ui/title';
