+++ src/features/school/study-with-us/study-with-us.tsx
@@ -1,7 +1,7 @@
 import { studyOptions } from './constants';
 import { OptionItem } from '@/app/components';
-import image from '@/assets/rs-school.webp';
 import Image from '@/features/image';
+import image from '@/shared/assets/rs-school.webp';
 import { Paragraph } from '@/shared/ui/paragraph';
 import { SectionLabel } from '@/shared/ui/section-label';
 import { Subtitle } from '@/shared/ui/subtitle';
