+++ src/features/trainers/angular.data.ts
@@ -1,11 +1,11 @@
 import type { Trainer } from './trainers.types';
-import aSacca from '@/assets/mentors/a-sacca.webp';
-import aSerhiyenia from '@/assets/mentors/a-serhiyenia.webp';
-import dKohut from '@/assets/mentors/d-kohut.webp';
-import kBritsyn from '@/assets/mentors/k-britsyn.webp';
-import nLoginova from '@/assets/mentors/n-loginova.webp';
-import oDuleba from '@/assets/mentors/o-duleba.webp';
-import rSaltykov from '@/assets/mentors/r-saltykov.webp';
+import aSacca from '@/shared/assets/mentors/a-sacca.webp';
+import aSerhiyenia from '@/shared/assets/mentors/a-serhiyenia.webp';
+import dKohut from '@/shared/assets/mentors/d-kohut.webp';
+import kBritsyn from '@/shared/assets/mentors/k-britsyn.webp';
+import nLoginova from '@/shared/assets/mentors/n-loginova.webp';
+import oDuleba from '@/shared/assets/mentors/o-duleba.webp';
+import rSaltykov from '@/shared/assets/mentors/r-saltykov.webp';
 
 export const angular: Trainer[] = [
   {
