+++ src/features/trainers/awsFundamentals.data.ts
@@ -1,6 +1,6 @@
 import type { Trainer } from './trainers.types';
-import awsImg2 from '@/assets/mentors/v-antonau.webp';
-import awsImg1 from '@/assets/mentors/v-kavaliou.webp';
+import awsImg2 from '@/shared/assets/mentors/v-antonau.webp';
+import awsImg1 from '@/shared/assets/mentors/v-kavaliou.webp';
 
 export const awsFundamentals: Trainer[] = [
   {
