+++ src/icons/angular-icon.tsx
@@ -1,5 +1,5 @@
-import angular from '@/assets/svg/angular-icon.svg';
 import Image from '@/features/image';
+import angular from '@/shared/assets/svg/angular-icon.svg';
 
 export const AngularIcon = () => {
   return <Image src={angular} alt="angular icon" />;
