+++ src/icons/arrow-icon.tsx
@@ -1,5 +1,5 @@
-import arrow from '@/assets/svg/arrow.svg';
 import Image from '@/features/image';
+import arrow from '@/shared/assets/svg/arrow.svg';
 
 export const ArrowIcon = ({ size = '24px' }: { size?: string }) => {
   return <Image src={arrow} alt="" height={size} width={size} />;
