+++ src/icons/aws.tsx
@@ -1,5 +1,5 @@
-import aws from '@/assets/svg/aws.svg';
 import Image from '@/features/image';
+import aws from '@/shared/assets/svg/aws.svg';
 
 export const AwsLogo = () => {
   return <Image src={aws} alt="AWS icon" />;
