+++ src/icons/dropdown-arrow.tsx
@@ -1,5 +1,5 @@
-import arrow from '@/assets/svg/dropdown-arrow.svg';
 import Image from '@/features/image';
+import arrow from '@/shared/assets/svg/dropdown-arrow.svg';
 
 export const DropdownArrow = () => {
   return <Image src={arrow} alt="dropdown-arrow" aria-label="dropdown-arrow" />;
