+++ src/icons/email.tsx
@@ -1,5 +1,5 @@
-import email from '@/assets/svg/email.svg';
 import Image from '@/features/image';
+import email from '@/shared/assets/svg/email.svg';
 
 export const EmailIcon = () => {
   return <Image src={email} alt="email icon" />;
