+++ src/icons/epam.tsx
@@ -1,5 +1,5 @@
-import epam from '@/assets/svg/epam.svg';
 import Image from '@/features/image';
+import epam from '@/shared/assets/svg/epam.svg';
 
 export const EpamLogo = () => {
   return <Image src={epam} alt="epam" />;
