+++ src/icons/facebook.tsx
@@ -1,5 +1,5 @@
-import facebook from '@/assets/svg/facebook.svg';
 import Image from '@/features/image';
+import facebook from '@/shared/assets/svg/facebook.svg';
 
 export const FacebookIcon = () => {
   return (
