+++ src/icons/javascript-icon.tsx
@@ -1,5 +1,5 @@
-import javascript from '@/assets/icons/js-secondary.webp';
 import Image from '@/features/image';
+import javascript from '@/shared/assets/icons/js-secondary.webp';
 
 export const JavascriptIcon = () => {
   return <Image src={javascript} alt="JavaScript icon" />;
