+++ src/icons/jetbrains.tsx
@@ -1,5 +1,5 @@
-import jetbrains from '@/assets/svg/jetbrains.svg';
 import Image from '@/features/image';
+import jetbrains from '@/shared/assets/svg/jetbrains.svg';
 
 export const JetBrainsLogo = () => {
   return <Image src={jetbrains} alt="jetbrains icon" />;
