+++ src/icons/nodejs-icon.tsx
@@ -1,5 +1,5 @@
-import nodeJS from '@/assets/svg/nodeJS.svg';
 import Image from '@/features/image';
+import nodeJS from '@/shared/assets/svg/nodeJS.svg';
 
 export const NodeJsIcon = () => {
   return <Image src={nodeJS} alt="nodeJS icon" />;
