+++ src/icons/react-icon.tsx
@@ -1,5 +1,5 @@
-import react from '@/assets/svg/react.svg';
 import Image from '@/features/image';
+import react from '@/shared/assets/svg/react.svg';
 
 export const ReactIcon = () => {
   return <Image src={react} alt="react icon" />;
