+++ src/icons/rs-banner.tsx
@@ -1,5 +1,5 @@
-import banner from '@/assets/svg/RsBanner.svg';
 import Image from '@/features/image';
+import banner from '@/shared/assets/svg/RsBanner.svg';
 
 export const RsBanner = () => {
   return <Image data-testid="rs-banner" src={banner} alt="RsBanner Icon" />;
