+++ src/icons/teach-It-forward-icon.tsx
@@ -1,5 +1,5 @@
-import teachItForwardIcon from '@/assets/svg/teachItForwardIcon.svg';
 import Image from '@/features/image';
+import teachItForwardIcon from '@/shared/assets/svg/teachItForwardIcon.svg';
 
 export const TeachItForwardIcon = () => {
   return <Image src={teachItForwardIcon} alt="Teach It Forward Icon" />;
