+++ src/icons/text-link-icon.tsx
@@ -1,5 +1,5 @@
-import textLinkIcon from '@/assets/svg/text-link.svg';
 import Image from '@/features/image';
+import textLinkIcon from '@/shared/assets/svg/text-link.svg';
 
 export const TextLinkIcon = () => {
   return <Image src={textLinkIcon} width="25px" height="25px" alt="external link icon" />;
